fix: recover non-compiling contract to a clean library + WASM build
The tipjar crate had never compiled (a parse error masked ~1500 errors
from incompatibly-merged feature PRs). Restore it to a clean
`cargo build` and `wasm32v1-none` release build.

Root-cause fixes:
- export=false on oversized DataKey/TipJarError/CreditError (>50-case
  Soroban spec limit)
- restore 21 lost storage sub-key enums + missing DataKey variants
- define InsuranceError; add ~95 missing error variants; renumber
  duplicate CreditError discriminants
- implement missing check_circuit_breaker/trigger_breaker/require_admin
- wire 3 undeclared modules; fix corrupted DataKey(Key()) placeholders
- error-type unification via From impls + .into(); Hash<32>->to_bytes();
  numeric casts; moved-value clones; import fixes
- relax deny(missing_docs) to warn

Also adds the address whitelist/blacklist feature (access_lists module).

+++ contracts/tipjar/src/strategy/execution.rs
@@ -93,11 +93,11 @@ pub fn execute_allocations(
     let now = env.ledger().timestamp();
 
     for (i, target) in strategy.allocations.iter().enumerate() {
-        if i >= distribution.len() {
+        if i as u32 >= distribution.len() {
             break;
         }
 
-        let amount = distribution.get(i).ok_or(ExecutionError::ExecutionFailed)?;
+        let amount = distribution.get(i as u32).ok_or(ExecutionError::ExecutionFailed)?;
 
         if amount <= 0 {
             continue;
